\name{dummy.code.lab}
\alias{dummy.code.lab}
\title{Same as psych::dummy.code but with labels}
\usage{
dummy.code.lab(x)
}
\arguments{
  \item{x}{nominal dataset}
}
\value{
labelled dataset
}
\description{
Same as psych::dummy.code but with labels
}
\examples{
Here are some examples
}
\seealso{
Other experimental: \code{\link{findr}};
\code{\link{ggheat}}; \code{\link{loopr2}};
\code{\link{loopr}}; \code{\link{omnicor}};
\code{\link{outerF}}; \code{\link{pseudocor}};
\code{\link{sigHeat}}; \code{\link{writeprop}};
\code{\link{xt}}
}

