\name{findlabs}
\alias{findlabs}
\title{Given a vector of variablenames, returns the variable labels in df
Useful for input into sankplot}
\usage{
findlabs(var, df)
}
\arguments{
  \item{var}{}

  \item{df}{}
}
\description{
Given a vector of variablenames, returns the variable
labels in df Useful for input into sankplot
}

