\name{loopr}
\alias{loopr}
\title{Loops through one or two datasets}
\usage{
loopr(xdat, ydat = NULL, fun = oplot, duplicatesOK = TRUE,
  printnote = F, ...)
}
\arguments{
  \item{xdat}{dataset}

  \item{ydat}{if given, a second dataset}

  \item{fun}{function e.g. correlation to relate the
  datasets}

  \item{extra}{named vector of arguments to fun e.g.}

  \item{extra}{named vector of arguments to fun e.g.}
}
\value{
Some stuff
}
\description{
Loops through one or two datasets
}
\examples{
Here are some examples
}
\seealso{
Other experimental: \code{\link{dummy.code.lab}};
\code{\link{findr}}; \code{\link{ggheat}};
\code{\link{loopr2}}; \code{\link{omnicor}};
\code{\link{otrans_inner}}; \code{\link{outerF}};
\code{\link{pseudocor}}; \code{\link{sigHeat}};
\code{\link{writeprop}}; \code{\link{xt}}
}

