\name{loopr1}
\alias{loopr1}
\title{Loops through one or two datasets}
\usage{
loopr1(xdat, fun = oplot, extra = NULL)
}
\arguments{
  \item{xdat}{dataset}

  \item{ydat}{if given, a second dataset}

  \item{fun}{function e.g. correlation to relate the
  datasets}

  \item{extra}{named vector of arguments to fun e.g.}
}
\value{
Some stuff
}
\description{
Loops through one or two datasets
}
\examples{
Here are some examples
}
\seealso{
Other experimental: \code{\link{dummy.code.lab}};
\code{\link{ggheat}}; \code{\link{omnicor}};
\code{\link{outerF}}; \code{\link{pseudocor}};
\code{\link{sigHeat}}; \code{\link{xt}}
}

