\name{loopr2}
\alias{loopr2}
\title{Loops through two lists and returns a  list of their pairs with no duplicates and no identical pairs. More general than loopr because x and y dont have to be dataframes}
\usage{
loopr2(x, y)
}
\arguments{
  \item{x}{list or vector}

  \item{y}{another}
}
\value{
Some stuff
}
\description{
Loops through two lists and returns a list of their pairs
with no duplicates and no identical pairs. More general
than loopr because x and y dont have to be dataframes
}
\examples{
Here are some examples
}
\seealso{
Other experimental: \code{\link{dummy.code.lab}};
\code{\link{findr}}; \code{\link{ggheat}};
\code{\link{loopr}}; \code{\link{omnicor}};
\code{\link{outerF}}; \code{\link{pseudocor}};
\code{\link{sigHeat}}; \code{\link{writeprop}};
\code{\link{xt}}
}

