\name{multimelt}
\alias{multimelt}
\title{PROBABLY OBSOLETE COS INCLUDED IN OTRANS Breaks out multi-choice dfs or combos of a df and a vector
Useful for input into sankplot}
\usage{
multimelt(a, b, nochars = xc("No no"), uselabs = T)
}
\arguments{
  \item{x}{}

  \item{y}{}
}
\description{
PROBABLY OBSOLETE COS INCLUDED IN OTRANS Breaks out
multi-choice dfs or combos of a df and a vector Useful for
input into sankplot
}

