\name{otrans_inner}
\alias{otrans_inner}
\title{doesnt touch xx or yy if they are just vectors. otherwise processes other structures into vectors.}
\usage{
otrans_inner(xx, yy = NULL, zz = NULL, xfilter = NULL, yfilter = NULL,
  mosaic = mosaic, usemultichoicelabs = F, table = F, reorder.x = F,
  reorder.y = F, tableout = F, msg = "")
}
\arguments{
  \item{xx}{sth}
}
\value{
Some stuff
}
\description{
doesnt touch xx or yy if they are just vectors. otherwise
processes other structures into vectors.
}
\examples{
Here are some examples
}
\seealso{
Other experimental: \code{\link{dummy.code.lab}};
\code{\link{findr}}; \code{\link{ggheat}};
\code{\link{loopr2}}; \code{\link{loopr}};
\code{\link{omnicor}}; \code{\link{outerF}};
\code{\link{pseudocor}}; \code{\link{sigHeat}};
\code{\link{writeprop}}; \code{\link{xt}}
}

