\name{outerF}
\alias{outerF}
\title{Applies function fun to all combinations of the variables in xdat and ydat}
\usage{
outerF(xdat, ydat = xdat, fun, ...)
}
\arguments{
  \item{xdat}{dataset}

  \item{ydat}{second dataset}
}
\value{
A matrix
}
\description{
Applies function fun to all combinations of the variables
in xdat and ydat
}
\examples{
Here are some examples
}
\seealso{
Other experimental: \code{\link{dummy.code.lab}};
\code{\link{findr}}; \code{\link{ggheat}};
\code{\link{loopr2}}; \code{\link{loopr}};
\code{\link{omnicor}}; \code{\link{otrans_inner}};
\code{\link{pseudocor}}; \code{\link{sigHeat}};
\code{\link{writeprop}}; \code{\link{xt}}
}

