\name{pseudocor}
\alias{pseudocor}
\title{Applies \code{otest} to each pair of vectors from two datasets, or within a single dataset}
\usage{
pseudocor(xdat, ydat = xdat)
}
\arguments{
  \item{xdat}{dataset}

  \item{ydat}{if given, a second dataset}
}
\value{
list of 3 matrices giving p-values, estimates and Ns
}
\description{
Applies \code{otest} to each pair of vectors from two
datasets, or within a single dataset
}
\examples{
Here are some examples
}
\seealso{
Other experimental: \code{\link{dummy.code.lab}};
\code{\link{findr}}; \code{\link{ggheat}};
\code{\link{loopr2}}; \code{\link{loopr}};
\code{\link{omnicor}}; \code{\link{otrans_inner}};
\code{\link{outerF}}; \code{\link{sigHeat}};
\code{\link{writeprop}}; \code{\link{xt}}
}

