\name{wplot}
\alias{wplot}
\title{Just a wrapper for wordcloud
separator element returns a}
\usage{
wplot(stri, cap = "", min.freq = 2, remove = NULL)
}
\arguments{
  \item{stri}{A string, in quotes}

  \item{sepp}{A string, usually a single character, to
  separate the elements of \code{stri}.}
}
\value{
A vector of strings.
}
\description{
Just a wrapper for wordcloud separator element returns a
}
\examples{
xc("red green blue")
xc("Croatia;Serbia;Bosnia and Herzegovina",sepp=";")
}
\seealso{
Other utility functions: \code{\link{xc}}
}

