\name{writeprop}
\alias{writeprop}
\title{ofix also creates a properties object as well writing a corresponding properties file
but although you can edit the properties object you have to write it to the file to ensure reproducibility.}
\usage{
writeprop(object = prop.data, file = "prop.data")
}
\arguments{
  \item{dat}{dataset}

  \item{att}{the attribute}

  \item{equ}{the value of the attribute}
}
\value{
A dataset
}
\description{
from the attribute
}
\examples{
Here are some examples
}
\seealso{
Other experimental: \code{\link{dummy.code.lab}};
\code{\link{findr}}; \code{\link{ggheat}};
\code{\link{loopr2}}; \code{\link{loopr}};
\code{\link{omnicor}}; \code{\link{otrans_inner}};
\code{\link{outerF}}; \code{\link{pseudocor}};
\code{\link{sigHeat}}; \code{\link{xt}}
}

