\name{xt}
\alias{xt}
\title{Wrapper for xtabs}
\usage{
xt(x, cap = "", typ = "html", cn = F, addmargins = 0, label = NULL,
  digits = 0, ...)
}
\arguments{
  \item{x}{a table}

  \item{cap}{the table caption}

  \item{typ}{type of output, default is html but can also
  be latex}

  \item{cn}{whether to show colnames}
}
\value{
A matrix
}
\description{
Wrapper for xtabs
}
\examples{
Here are some examples
}
\seealso{
Other experimental: \code{\link{dummy.code.lab}};
\code{\link{findr}}; \code{\link{ggheat}};
\code{\link{loopr2}}; \code{\link{loopr}};
\code{\link{omnicor}}; \code{\link{otrans_inner}};
\code{\link{outerF}}; \code{\link{pseudocor}};
\code{\link{sigHeat}}; \code{\link{writeprop}}
}

