\name{ggheat}
\alias{ggheat}
\title{Produces a ggplot heatmap}
\usage{
ggheat(xdat, ydat = xdat)
}
\arguments{
  \item{xdat}{dataset}

  \item{ydat}{if given, a second dataset}

  \item{fun}{function e.g. correlation to relate the
  datasets}

  \item{ydat}{if given, a second dataset}

  \item{ydat}{if given, a second dataset}
}
\value{
A heatmap.2 plot
}
\description{
Produces a ggplot heatmap
}
\examples{
Here are some examples
}
\seealso{
Other experimental: \code{\link{dummy.code.lab}};
\code{\link{findr}}; \code{\link{loopr2}};
\code{\link{loopr}}; \code{\link{omnicor}};
\code{\link{otrans_inner}}; \code{\link{outerF}};
\code{\link{pseudocor}}; \code{\link{sigHeat}};
\code{\link{writeprop}}; \code{\link{xt}}
}

