\name{xmb}
\alias{xmb}
\title{More robust test of whether something is not blank, null etc.
true if missing or null or y, otherwise false. NOTE IT GIVES F IF IT IS ANY DATA FRAME, EVEN AN EMPTY ONE}
\usage{
xmb(x, y = "")
}
\description{
More robust test of whether something is not blank, null
etc. true if missing or null or y, otherwise false. NOTE IT
GIVES F IF IT IS ANY DATA FRAME, EVEN AN EMPTY ONE
}

