\name{xt}
\alias{xt}
\title{Wrapper for xtabs}
\usage{
xt(x, cap = "", typ = "html", cn = F, ...)
}
\arguments{
  \item{x}{a table}
}
\value{
A matrix
}
\description{
Wrapper for xtabs
}
\examples{
Here are some examples
}
\seealso{
Other experimental: \code{\link{dummy.code.lab}};
\code{\link{findr}}; \code{\link{ggheat}};
\code{\link{loopr2}}; \code{\link{loopr}};
\code{\link{omnicor}}; \code{\link{outerF}};
\code{\link{pseudocor}}; \code{\link{sigHeat}};
\code{\link{writeprop}}
}

